getBiometriType from deviceInfoService
