+++ packages/core-mobile/app/new/features/onboarding/components/CreatePin.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useRef } from 'react'
+import React, { useCallback, useRef, useEffect, useState } from 'react'
 import { useFocusEffect } from 'expo-router'
 import BlurredBarsContentLayout from 'common/components/BlurredBarsContentLayout'
 import {
@@ -14,7 +14,9 @@ import { useCreatePin } from 'features/onboarding/hooks/useCreatePin'
 import { InteractionManager } from 'react-native'
 import ScreenHeader from 'common/components/ScreenHeader'
 import { KeyboardAvoidingView } from 'common/components/KeyboardAvoidingView'
-import { useDeviceInfoContext } from 'common/contexts/DeviceInfoProvider'
+import DeviceInfoService, {
+  BiometricType
+} from 'services/deviceInfo/DeviceInfoService'
 
 export const CreatePin = ({
   useBiometrics,
@@ -26,8 +28,9 @@ export const CreatePin = ({
   onEnteredValidPin: (validPin: string) => void
 }): React.JSX.Element => {
   const ref = useRef<PinInputActions>(null)
-  const { bioType } = useDeviceInfoContext()
-
+  const [biometricType, setBiometricType] = useState<BiometricType>(
+    BiometricType.NONE
+  )
   const {
     onEnterChosenPin,
     onEnterConfirmedPin,
@@ -45,6 +48,14 @@ export const CreatePin = ({
       })
     }
   })
+
+  useEffect(() => {
+    const getBiometryType = async (): Promise<void> => {
+      const type = await DeviceInfoService.getBiometricType()
+      setBiometricType(type)
+    }
+    getBiometryType()
+  }, [])
 
   useFocusEffect(
     useCallback(() => {
@@ -108,7 +119,7 @@ export const CreatePin = ({
               <GroupList
                 data={[
                   {
-                    title: `Unlock with ${bioType}`,
+                    title: `Unlock with ${biometricType}`,
                     accessory: (
                       <Toggle
                         onValueChange={setUseBiometrics}
